chore: added update movie

+++ app/Controllers/Http/MovieController.js
@@ -79,16 +79,78 @@ class MovieController {
 
   async editMovie({request,response,user,params}){
 
-    const movie = await Movie.query().whereHas("users",(b)=>{b.where("user_id",user.id)}).with("media").where("id",57).first()
-
-    fs.unlink(`./public/${movie.$relations.media.path}`,(err) => {
-      if (err) {
-        console.log("failed to delete local image:" + err);
-      } else {
-        console.log('successfully deleted local image');
-      }
+    const allParams = sanitize(request.post(),{
+      duration:"to_int",
+      finished:"to_boolean"
     })
 
+    const validation = await validate(allParams,{
+      title:"required|max:50",
+      description:"required|max:254",
+      genre:"required",
+      duration:"required"
+    })
+
+    if(validation.fails()){
+      return response.badRequest({
+        _message:"Invalid params"
+      })
+    }
+
+    const movie = await Movie.query().whereHas("users",(b)=>{b.where("user_id",user.id)}).with("media").where("id",params.id).first()
+
+
+    if(!movie){
+      return response.notFound({
+        _message:"Movie does not exist"
+      })
+    }
+
+
+    const movieCover = request.file("cover_picture",{
+      types:["image"],
+      size:"2mb"
+    })
+    if(movieCover) {
+
+
+      const pictureName = `${Moment().format("YYYY-MM-DD-HH-mm-ss")}.jpg`
+      await movieCover.move(Helpers.publicPath("/resources/media"),{
+        name: `${pictureName}`
+      })
+
+      if(!movieCover.moved()){
+        return movieCover.error()
+      }
+
+      fs.unlink(`./public/${movie.$relations.media.path}`, (err) => {
+        if (err) {
+          console.log("failed to delete local image:" + err);
+        } else {
+          console.log('successfully deleted local image');
+        }
+      })
+
+      const media = await Media.query().where("movie_id",movie.id).first()
+
+      media.merge({
+        path: `resources/media/${pictureName}`
+      })
+
+      media.save()
+    }
+
+    movie.merge({
+      title:allParams.title,
+      description:allParams.description,
+      genre:allParams.genre,
+      duration:allParams.duration
+    })
+    movie.save()
+
+    await movie.users().pivotQuery().update({finished:allParams.finished})
+
+    return response.ok()
   }
 
   async deleteMovie({response, params,user}){
